🔒 fix(security): resolve XSS vulnerability in DNSLookup handler
This commit fixes a Cross-Site Scripting (XSS) vulnerability in the
DNSLookup handler where error messages from the DNS service were being
rendered directly into the HTML response without escaping.

Changes:
- Wrapped the error message in `html.EscapeString()` in `internal/handler/handlers.go`.
- Verified that other `c.HTML` calls in the same file already use proper escaping.
- Added a `DNSLookupFunc` hook in `internal/service/dns_hook.go` to support better testing/mocking of DNS lookups.
- Refactored `Lookup` method from `dns.go` to `dns_hook.go` to facilitate the test hook.
